requirements.txt
Flask==3.1.2 
gunicorn==23.0.0 
python-docx==1.2.0 
lxml==6.0.2 
textract==1.6.1 
